Move and rename 'su2functions' to 'su2meshutils'
