+++ ceasiompy/SU2MeshDef/su2meshdef.py
@@ -43,7 +43,7 @@ from cpacspy.cpacsfunctions import (
     open_tigl,
     open_tixi,
 )
-from ceasiompy.utils.su2functions import get_mesh_marker
+from ceasiompy.SU2Run.func.su2meshutils import get_mesh_marker
 from ceasiompy.utils.configfiles import ConfigFile
 from ceasiompy.utils.xpath import REF_XPATH, WINGS_XPATH, SU2_XPATH
