+++ ceasiompy/SU2Run/func/su2meshutils.py
@@ -90,4 +90,4 @@ if __name__ == "__main__":
 
     print("Nothing to execute!")
     print("You can use this module by importing:")
-    print("from ceasiompy.utils.su2functions import get_mesh_marker")
+    print("from ceasiompy.SU2Run.func.su2meshutils import get_mesh_marker")
